full-core
The full-core model is working in Fulcrum and looks good.

You shouldn't try to load any segmented full-core input files for viewing in Fulcrum since it will most likely crash your computer.

Unfortunately there are lost particles that are all in the positive y direction which indicates what I would consider a bug in KENO-VI. But I can change my model to accomadate for this bug.

Essentially each thirst of an assmebly containing the active regions and channels is arrayed together by axial regions into a parallelogram unit. This unit is the holed into the assembly unit where it does not touch the boundary. Unfortunately, this causes problems in KENO-VI with particles getting lost all in the positive y direction.

The previous full-core model that worked was almost identical but without the axial layers. So the solution is to instead make axial arrays of TRISO particles and then the fuel and channel units cover the full axial dimension in a single unit. This way the array filling the unit is only one level deep in the hole that goes into the assembly. It seems that having two levels of arrays is what generates the error. This however seems like an inefficient MC division  of units.