full-core that works in SCALE
there was an issue with SCALE not being able to run correctly if the core's array ID in the global unit was greater than the array IDs of other arrays in the model. This was the problem with my geometry not an actual problem with my geometry but a strange bug/feature in SCALE I was unaware of.